% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_genotypes.R
\name{format_genotypes}
\alias{format_genotypes}
\title{Transform genotype format to R/qtl2's}
\usage{
format_genotypes(geno_ag_KC)
}
\arguments{
\item{geno_ag_KC, }{genotype table}
}
\value{
the recoded genotypes
}
\description{
Transform genotype format to R/qtl2's
}
\examples{
# TODO: example
}
