% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_f_codes.R
\name{get_f_codes}
\alias{get_f_codes}
\title{Read crossing tables et generate random data to replace missing data}
\usage{
get_f_codes(file_f_codes)
}
\arguments{
\item{file_f_codes, }{path to crossing data}
}
\value{
funnel codes
}
\description{
Read crossing tables et generate random data to replace missing data
}
\examples{
# TODO: example
}
