% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeneize_gen.R
\name{homogeneize_gen}
\alias{homogeneize_gen}
\title{Get genotype files (founders and individuals) with identical loci}
\usage{
homogeneize_gen(file_founder, file_individuals)
}
\arguments{
\item{file_founder, }{string, path to founder file}

\item{file_individuals, }{string, path to individual files}
}
\value{
list of genotypes and loci
}
\description{
Get genotype files (founders and individuals) with identical loci
}
\examples{
# TODO: example
}
