% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2N.R
\name{L2N}
\alias{L2N}
\title{Transform letters into numbers}
\usage{
L2N(x)
}
\arguments{
\item{x, }{string}
}
\value{
characters
}
\description{
Transform letters into numbers
}
\examples{
L2N("N")
sapply(c("A", "C", "G"), L2N)

}
