% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_r2r.R
\name{match_r2r}
\alias{match_r2r}
\title{Get phenotypes, crossing and genotypes and match them}
\usage{
match_r2r(geno, pheno, fun_codes)
}
\arguments{
\item{geno, }{genotype table}

\item{pheno, }{phenotype table}

\item{fun_codes, }{"funnel" codes}
}
\value{
a list with three elements: the genotype table, 
the phenotype table and the funnel codes for the common
individuals between the three tables.
}
\description{
Get phenotypes, crossing and genotypes and match them
}
\examples{
# TODO: example
}
